Moving common functions to utils
